mt7620 SPI: Mark the SPI master as half duplex.
It appears that full duplex is not supported in this hardware. I tried
moving the rx and tx transfers to the same loop in rt2880_spi_write_read,
setting the SPICTL_STARTWR and SPICTL_STARTRD bits in the same write to
SPI_CTL, but only the read operation was performed, and it wrote 0x00.

Signed-off-by: Kevin Mehall <km@kevinmehall.net>